refactor: clear all rustc and clippy warnings tree-wide
Zero warnings across `cargo build` and `cargo clippy --all-targets`:

- drop dead code (SoftCounts::merge, DgqStore::profile, TurnState::sequences)
- gate test-only helpers behind #[cfg(test)]; #[allow(dead_code)] the parked
  KV-rewind and diagnostic scaffolding, each with a why-kept note
- rewrite needless_range_loop sites to iterators in the oracle/kernel-mirror
  math; #[allow] only where a loop indexes two arrays in lockstep
- fix the kernel_oracle_matrix! macro (non_snake_case module + unused imports),
  drop 28 unused `use super::*`, and misc borrow/lifetime/precision cleanups
- ServerEvent -> pub(crate); drop the no-op `pub use enc::*`

Behaviour-preserving: golden 8/8 byte-identical, smoketest 17/17, full suite
749 passing.

+++ src/dgq/block.rs
@@ -286,7 +286,7 @@ mod tests {
         let mut src = vec![0u8; out_dim * in_dim * 2];
         for row in 0..out_dim {
             for col in 0..in_dim {
-                let v = ((row * in_dim + col) as f32 * 0.01 - 0.5) as f32;
+                let v = (row * in_dim + col) as f32 * 0.01 - 0.5;
                 let bits = f32_to_bf16_bits(v).to_le_bytes();
                 let i = (row * in_dim + col) * 2;
                 src[i] = bits[0];
